refactor(frontend): 🎨 simplify Socket Context
Now all events are handled by the context, which makes things simpler
